requirements.txt
boto3
flask
moviepy 
google-cloud-speech
pydub
Speechrecognition
librosa